+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-08-24
+
+### Fixed
+
+- Fixed bug with build description textarea sizing inconsistencies.
+
 ## 2024-08-22
 
 ### Fixed